requirements.txt
# Core AI/ML Libraries
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.20
openai>=1.0.0
anthropic>=0.7.0

# Vector Database & Memory
chromadb>=0.4.22
pinecone-client>=3.0.0
weaviate-client>=3.25.0

# Web Framework & API
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database & Caching
redis>=5.0.0
pymongo>=4.6.0
sqlalchemy>=2.0.0

# Speech Processing
SpeechRecognition>=3.10.0
pyttsx3>=2.90
gtts>=2.5.0
pydub>=0.25.1
pyaudio>=0.2.14
whisper>=1.1.10

# Vision Processing
opencv-python>=4.9.0
pillow>=10.2.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0

# NLP & Text Processing
spacy>=3.7.0
transformers>=4.37.0
sentence-transformers>=2.3.0
nltk>=3.8.1

# Google Cloud & APIs
google-cloud-speech>=2.24.0
google-cloud-texttospeech>=2.16.0
google-cloud-vision>=3.5.0
google-cloud-storage>=2.14.0
google-api-python-client>=2.115.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0

# Web Scraping & Search
beautifulsoup4>=4.12.0
selenium>=4.17.0
requests>=2.31.0
httpx>=0.26.0
playwright>=1.41.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
polars>=0.20.0

# Async & Task Management
asyncio>=3.4.3
aiohttp>=3.9.0
celery>=5.3.0

# Message Queue
pika>=1.3.0
kafka-python>=2.0.0

# Monitoring & Logging
prometheus-client>=0.19.0
python-json-logger>=2.0.7
sentry-sdk>=1.40.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
toml>=0.10.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0

# Development Tools
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0
ipython>=8.20.0
jupyter>=1.0.0

# Deployment
gunicorn>=21.2.0
docker>=7.0.0
kubernetes>=28.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
schedule>=1.2.0
croniter>=2.0.0
tenacity>=8.2.0
